Add IterativeSaver class
